+++ src/poseidon2/params.rs
@@ -35,7 +35,7 @@ impl<E: Engine, const RATE: usize, const WIDTH: usize> Default for Poseidon2Para
         // Number of rounds from the original Poseidon2 implementation
         // https://github.com/HorizenLabs/poseidon2
         let full_rounds = 8;
-        let partial_rounds = 56;
+        let partial_rounds = 33;
 
         let mut params = InnerHashParameters::<E, RATE, WIDTH>::new(security_level, full_rounds, partial_rounds);
